requirements.txt
# HYDRA V2 - Dependencies
# Para Railway/Render deployment

# FastAPI + Server
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Database (Supabase)
supabase>=2.0.0

# Telegram Notifications
pyTelegramBotAPI>=4.14.0

# Google AI (Gemini)
google-generativeai>=0.3.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.9.0

# Environment
python-dotenv>=1.0.0

# RSS Feeds
feedparser>=6.0.0

# WebSockets
websockets>=12.0

# Pydantic Settings
pydantic-settings>=2.0.0
